Create model classes for all entities as shown in the ERD diagram and Set up associations between models.

+++ db/migrate/20221208183803_create_comments.rb
@@ -1,8 +1,8 @@
 class CreateComments < ActiveRecord::Migration[7.0]
   def change
     create_table :comments do |t|
-      t.integer :author_id
-      t.integer :post_id
+      t.bigint :author_id
+      t.bigint :post_id
       t.string :text
 
       t.timestamps
